Seperated test scenes to be self-contained
