+++ pyscenes/examples/title_screen.py
@@ -1,10 +1,4 @@
-# Centralized Scene Logic inspired by https://nerdparadise.com/programming/pygame/part7
-# this is where the game itself comes to life
-# the user should make their own scenes within this class
-#   expecting that the runner will take care of the game loop itself
-#   so the user can focus on making scenes using the methods available in
-#   PyScenes
-
+#Self-contained example of using PyScenes for a simple title screen
 import pyscenes.pyscenes as pyscenes
 from pyscenes.base_scene import BaseScene
 
@@ -13,13 +7,13 @@ SCREEN_WIDTH = 800
 SCREEN_HEIGHT = 600
 FPS = 30
 
-# Title Scene
-
-
 class TitleScene(BaseScene):
-    def __init__(self, display, test=False):
+    '''
+    TitleScene example class.
+    '''
+    def __init__(self, game, test=False):
         print("Initializing TitleScene...")
-        self.display = display
+        self.game = game
         if test:
             self.test = test
             self.render_count = 0
@@ -29,7 +23,7 @@ class TitleScene(BaseScene):
 
     def setup(self):
         print("Setting up TitleScene...")
-        self.display.background.setBackgroundImage("assets/title_screen.jpg")
+        self.game.display.background.setBackgroundImage("assets/title_screen.jpg")
 
     def process_input(self, events, pressed_keys):
         print("process")
@@ -49,11 +43,11 @@ class TitleScene(BaseScene):
         print("cleanup")
 
 
-# def main():
-#     instance = pyscenes.Game(SCREEN_WIDTH, SCREEN_HEIGHT, FPS)
-#     # pass a Scene object here to start the game
-#     instance.run_game(TitleScene(instance.display, True))
+def main():
+    instance = pyscenes.Game(SCREEN_WIDTH, SCREEN_HEIGHT, FPS)
+    # pass a Scene object here to start the game
+    instance.run_game(TitleScene(instance.display, True))
 
-# TODO: only Run this if file executed as script
-# if __name__ == "__main__":
-#     main()
+TODO: only Run this if file executed as script
+if __name__ == "__main__":
+    main()
