+++ tests/test_runner.py
@@ -1,6 +1,6 @@
 import pyscenes.pyscenes as pyscenes
-from pyscenes.base_scene import BaseScene
 from pyscenes.examples.title_screen import TitleScene
+from counter_scene import CounterScene
 import os
 import pytest
 import time
@@ -9,54 +9,10 @@ os.environ["SDL_VIDEODRIVER"] = "dummy"
 os.environ["SDL_AUDIODRIVER"] = "dummy"
 
 
-# Game constants should be defined in scenes
+# Define Game constants here for same size
 SCREEN_WIDTH = 800
 SCREEN_HEIGHT = 600
 FPS = 30
-
-
-class CounterScene(BaseScene):
-    '''
-    CounterScene is a minimum implementation of a Scene Object
-    This scene is setup to make sure each method inside the scene is run the appropriate amount of time according to the FPS.
-    '''
-
-    def __init__(self, test_duration):
-        self.test_duration = test_duration * FPS
-        self.setup_count = 0
-        self.process_input_count = 0
-        self.update_count = 0
-        self.render_count = 0
-        self.cleanup_count = 0
-        super().__init__()
-
-    def setup(self):
-        self.setup_count += 1
-        print("setup called", self.setup_count, "times")
-
-    def process_input(self, events, pressed_keys):
-        self.process_input_count += 1
-
-    def update(self):
-        self.update_count += 1
-
-    def render(self, screen):
-        self.render_count += 1
-
-        if self.render_count == self.test_duration:
-            super().terminate()
-
-    def cleanup(self):
-        self.cleanup_count += 1
-        print("setup called", self.cleanup_count, "times")
-
-        print("setup test")
-        assert self.setup_count == 1
-        assert self.process_input_count == self.test_duration
-        assert self.update_count == self.test_duration
-        assert self.render_count == self.test_duration
-        print("cleanup test")
-        assert self.cleanup_count == 1
 
 
 @pytest.fixture
@@ -117,4 +73,4 @@ def test_counter_scene(game_instance):
     '''
     See CounterScene above.
     '''
-    game_instance.run_game(CounterScene(2))
+    game_instance.run_game(CounterScene(game_instance, 2))
